requirements.txt
espnet~=202301
numpy~=1.24.2
six~=1.16.0
scipy~=1.10.1
torch~=1.13.1
fastapi~=0.92.0